test: replace inline tests with proper success/failure/edge test structure for lambda and let
