+++ crates/core/src/eval/functions/logical/lambda/tests/mod.rs
@@ -1,0 +1,3 @@
+mod success;
+mod failure;
+mod edge;
